requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
httpx==0.27.2
psycopg2-binary==2.9.9